adjusting to checklist
