+++ app.js
@@ -3,11 +3,13 @@ const path = require('path');
 const bodyParser = require('body-parser');
 const mainRoute = require('./routes/main');
 const gamesRouter = require('./routes/games'); 
+const cors = require('./middlewares/cors')
 
 const PORT = 3000;
 const app = express();
 
 app.use(
+    cors,
     bodyParser.json(),
     express.static(path.join(__dirname, 'public')),
     mainRoute,
