+++ routes/games.js
@@ -1,7 +1,7 @@
      // routes/games.js
      const gamesRouter = require("express").Router(); // Создали роутер
-     const { sendAllGames, deleteGame, addGameController } = require("../controllers/games")
-     const { getAllGames } = require("../middlewares/games")
+     const { sendAllGames } = require("../controllers/games")
+     const { getAllGames, deleteGame, addGameController } = require("../middlewares/games")
      
      gamesRouter.get("/games", getAllGames, sendAllGames)
      gamesRouter.delete("/games/:id", getAllGames, deleteGame);
